Groq -API-/requirements.txt
Flask
groq
python-dotenv
python-docx
PyMuPDF
Pandas
openpyxl
tabulate